Without this patch, the folder settings page lives at /folders/$folderName, which prevents using that URL for a folder index page (collection index with data grid). The org and project patterns both use a /settings subroute, but folders did not.
This patch moves FolderDetailPage to /folders/$folderName/settings to match the URL
structure used by organizations (/orgs/$orgName/settings) and projects
(/projects/$projectName/settings).  The bare /folders/$folderName path now redirects
to /folders/$folderName/settings, following the same pattern as the project layout.

Changes:
- Created frontend/src/routes/_authenticated/folders/$folderName/settings/index.tsx
  with FolderDetailPage moved from the old index.tsx
- Replaced folders/$folderName/index.tsx with a Navigate redirect to settings
- Updated breadcrumbs in the settings page to include a Settings segment
- Updated the templates page breadcrumb link to point to settings
- Updated test imports to reference the new settings module path
- Fixed test assertion for folder slug (now appears in both breadcrumb and slug field)

Result: make test passes (688 UI tests, all Go tests). The folder settings page
renders at /folders/$folderName/settings and bare /folders/$folderName redirects
there. Org-scoped redirect at /orgs/$orgName/folders/$folderName continues to work
since it redirects to /folders/$folderName which redirects to settings.

Refs: https://github.com/holos-run/holos-console/issues/750
